♻️ Make local-config annotation specific
kpt only does a function at once, and honors the removal
of the `config.kuberntes.io/local-config` marked resources.

In consequence, we cannot use `config.kuberntes.io/local-config`
anymore to inject resources. We replace it with
 `config.kaweezle.com/local-config`.

`prune-local` is not a workaround anymore  but becomes
a feature. The README needs to be adapted in consequence.