🐛 ✨ Introduce new annotation to cleanup annotations
After transformation, we were deleting all internal annotations.
When using krmfnbuiltin in the context of a kustomize build, we
don't want those annotation to disappear. For instance, this would
prevent hash generation for secrets and configmaps.

But in the context of a single function execution, we want to keep
it. This is the reason we have added a specific annotation for that.